# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MemoryBridge Server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/memory/query": {
      "get": {
        "tags": [
          "记忆"
        ],
        "summary": "Query Memories",
        "description": "向量相似度检索：返回与 query 最相关的记忆条目。\n\n用途：调试面板展示 RAG 检索结果，验证记忆是否被正确存储和检索。",
        "operationId": "query_memories_api_memory_query_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "用户 ID",
              "title": "User Id"
            },
            "description": "用户 ID"
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "description": "检索文本",
              "title": "Query"
            },
            "description": "检索文本"
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "返回条数，0 表示用默认配置",
              "default": 0,
              "title": "Top K"
            },
            "description": "返回条数，0 表示用默认配置"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoryResponse"
                  },
                  "title": "Response Query Memories Api Memory Query Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "importance": {
            "type": "integer",
            "title": "Importance",
            "default": 1
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "MemoryResponse",
        "description": "记忆检索结果。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}